feat: break db.php up into autoloadable classes
